oidcccl: move TestMain to main_test.go
... as is customary. main_test.go comes with the advantage that it can
be in the _test package, which I'll use later to have it import all
"ccl".

Release note: None
